@@ -1,6 +1,6 @@
 {
   "name": "finsemble-seed",
-  "version": "6.3.0",
+  "version": "6.4.0-beta.2",
   "description": "",
   "main": "index.js",
   "scripts": {
@@ -29,12 +29,12 @@
   "author": "Cosaic",
   "license": "SEE LICENSE IN LICENSE",
   "optionalDependencies": {
-    "@finsemble/finsemble-cli": "6.3.0"
+    "@finsemble/finsemble-cli": "6.4.0-beta.2"
   },
   "dependencies": {
-    "@finsemble/finsemble-core": "6.3.0",
-    "@finsemble/finsemble-electron-adapter": "6.3.0",
-    "@finsemble/finsemble-ui": "6.3.0",
+    "@finsemble/finsemble-core": "6.4.0-beta.2",
+    "@finsemble/finsemble-electron-adapter": "6.4.0-beta.2",
+    "@finsemble/finsemble-ui": "6.4.0-beta.2",
     "@svgr/webpack": "^5.5.0",
     "async": "3.2.1",
     "compression": "1.7.4",
@@ -45,24 +45,24 @@
     "lodash": "4.17.21"
   },
   "devDependencies": {
-    "@babel/core": "7.15.8",
-    "@babel/plugin-proposal-class-properties": "7.14.5",
-    "@babel/plugin-proposal-decorators": "7.15.8",
-    "@babel/plugin-proposal-export-default-from": "7.14.5",
-    "@babel/plugin-transform-modules-commonjs": "7.15.4",
-    "@babel/plugin-transform-runtime": "7.15.8",
-    "@babel/preset-env": "7.15.6",
-    "@babel/preset-react": "7.14.5",
-    "@babel/runtime": "7.15.4",
-    "@types/react-dom": "17.0.8",
+    "@babel/core": "7.16.0",
+    "@babel/plugin-proposal-class-properties": "7.16.0",
+    "@babel/plugin-proposal-decorators": "7.16.4",
+    "@babel/plugin-proposal-export-default-from": "7.16.0",
+    "@babel/plugin-transform-modules-commonjs": "7.16.0",
+    "@babel/plugin-transform-runtime": "7.16.4",
+    "@babel/preset-env": "7.16.4",
+    "@babel/preset-react": "7.16.0",
+    "@babel/runtime": "7.16.3",
+    "@types/react-dom": "17.0.11",
     "@types/react-syntax-highlighter": "13.5.2",
-    "babel-loader": "8.2.2",
+    "babel-loader": "8.2.3",
     "babel-plugin-add-module-exports": "1.0.4",
     "case-sensitive-paths-webpack-plugin": "2.4.0",
     "chalk": "4.1.2",
-    "copy-webpack-plugin": "9.0.1",
+    "copy-webpack-plugin": "9.1.0",
     "cross-env": "7.0.3",
-    "css-loader": "3.5.3",
+    "css-loader": "6.5.1",
     "death": "1.1.0",
     "file-loader": "6.2.0",
     "gulp": "4.0.2",
@@ -70,17 +70,17 @@
     "improved-yarn-audit": "^3.0.0",
     "minimist": "1.2.5",
     "pretty-hrtime": "1.0.3",
-    "react-syntax-highlighter": "15.4.4",
+    "react-syntax-highlighter": "15.4.5",
     "rimraf": "3.0.2",
     "shelljs": "0.8.4",
     "source-map-loader": "3.0.0",
-    "style-loader": "3.3.0",
+    "style-loader": "3.3.1",
     "tree-kill": "1.2.2",
     "ts-loader": "9.2.6",
     "tslint": "6.1.3",
     "typescript": "4.3.5",
     "url-loader": "4.1.1",
-    "webpack": "5.59.1",
+    "webpack": "5.64.0",
     "webpack-bundle-analyzer": "4.4.2",
     "webpack-cli": "4.9.1",
     "webpack-glob-entries": "1.0.1",
